+++ mdsf/src/runners/deno.rs
@@ -1,5 +1,5 @@
 #[inline]
-pub fn setup_deno_run_command(package_name: &str, executable_name: &str) -> std::process::Command {
+pub fn setup_deno_run_command(package_name: &str, _executable_name: &str) -> std::process::Command {
     let mut cmd = std::process::Command::new("deno");
 
     cmd.arg("run");
@@ -39,6 +39,40 @@ mod test_deno {
     #[test_with::executable(deno)]
     #[test]
     fn it_works_with_executable_name() {
-        todo!()
+        let input = r#"model Pet {  name: string;  age: int32;kind: "dog" | "cat" | "fish";}
+"#;
+
+        let output = r#"model Pet {
+  name: string;
+  age: int32;
+  kind: "dog" | "cat" | "fish";
+}
+"#;
+
+        let file_ext = crate::filetype::get_file_extension("typespec");
+
+        let snippet =
+            crate::execution::setup_snippet(input, &file_ext).expect("it to create a snippet file");
+
+        let result = crate::execution::run_tools(
+            &[crate::runners::CommandType::Deno(
+                "@typespec/compiler",
+                "tsp",
+            )],
+            snippet.path(),
+            crate::tools::tsp_format::set_args,
+            crate::testing::DEFAULT_TEST_FORMATTER_TIMEOUT,
+            crate::tools::tsp_format::IS_STDIN,
+            crate::testing::DEFAULT_TEST_DEBUG_ENABLED,
+            &crate::config::MdsfConfigRunners {
+                deno: true,
+                ..Default::default()
+            },
+        )
+        .expect("it to be successful")
+        .1
+        .expect("it to be some");
+
+        assert_eq!(result, output);
     }
 }