+++ mdsf/src/tools/mojo_format.rs
@@ -15,6 +15,10 @@ pub fn set_args(
     cmd
 }
 
-pub const COMMANDS: [CommandType; 1] = [CommandType::Direct("mojo")];
+pub const COMMANDS: [CommandType; 3] = [
+    CommandType::Direct("mojo"),
+    CommandType::Uv("modular", "mojo"),
+    CommandType::Pipx("modular", "mojo"),
+];
 
 pub const IS_STDIN: bool = false;